Refactor_GeneralRules_Service
